review

+++ Job Portal Management System/Views/Company/CompanyEdit.aspx.cs
@@ -60,7 +60,7 @@ namespace Job_Portal_Management_System.Views.Company
             companymodel.CountryID = Convert.ToInt32(ddlCountry.SelectedValue);
             companymodel.Address = txtAddress.Text;
             companymodel.ContactPerson = txtContactPerson.Text;
-            companymodel.Mobile = Convert.ToInt32(txtMobile.Text);
+            companymodel.Mobile = Convert.ToInt64(txtMobile.Text);
             companymodel.Website = txtWebsite.Text;
             companymodel.Detail = txtDetail.Text;
             companymodel.UpdatedDate = DateTime.Now;
